fix(DmSynth): implement stop-gap solution to fix audio flicker when switching bands
This only really works for Gothic 1 and doesn't actually fix the root cause which is, that newly allocated TSFs will produce PCM that is drastically different from what was produced by the previous one, causing audio flicker.

In the future, a solution might be to fade out the audio to zero before the transition happens.