[#4] fix: improve handling of drum kit instruments
This patch fixes an issue where Gothic 2 requires drum kits to function correctly but Gothic 1 behaves incorrectly when a drum kit is allowed to be chosen.

In particular, Gothic 1 assigns the same performance channel (channel 7) to both a melodic and a drum instrument. The drum instrument in this case is a metronome ("Metronom") which is irrelevant for music playback. Previously, if instruments with the drum kit flag considered), we would always choose the drum kit instead of the melodic instrument for that channel, which is definitely incorrect.

It turns out, however, that ignoring drum kits is not the correct behaviour, since Gothic 2 requires them to be chosen (e.g. for `MO_DayStd.sgt`).

To get around this conundrum, we now simply act as if instruments with the same performance channel override each other, so that the last instrument in the sequence wins. This way, both soundtracks choose the right instruments, even if they are drum kits.